Created Good Map

+++ src/prefabs/Star.js
@@ -1,6 +1,7 @@
 class Star extends Entity {
     constructor(scene, x, y) {
         super(scene, x, y, "star", 10, 10);
+        console.log(scene)
         this.t = 0;
         this.bobSpeed = 0.05;
     }
@@ -8,9 +9,14 @@ class Star extends Entity {
     physicsUpdate() {
         this.t += 1;
 
+        this.pickup = true;
         this.getColliding().forEach(entity => {
+            if (!this.pickup) return;
+            this.pickup = false;
+
             this.scene.kroq.stars += 1;
             UI.instance.starText.text = "x" + this.scene.kroq.stars;            
+            this.scene.sound.add("starPickup").setVolume(0.1).play();
             this.delete();
             return;
         })
